+++ packages/astro/src/assets/fonts/implementations/css-renderer.ts
@@ -1,11 +1,13 @@
-import type { FontFaceData } from 'unifont';
 import type { CssRenderer } from '../definitions.js';
-import { generateFontFace } from '../utils.js';
+import { renderFontFace } from '../utils.js';
 
 // TODO: better impl
 export class PrettyCssRenderer implements CssRenderer {
-	generateFontFace(family: string, font: FontFaceData): string {
-		return generateFontFace(family, font);
+	generateFontFace(family: string, properties: Record<string, string | undefined>): string {
+		return renderFontFace({
+			'font-family': family,
+			...properties,
+		});
 	}
 	generateCssVariable(key: string, values: Array<string>): string {
 		return `:root { ${key}: ${values.join(', ')}; }`;
